added new enpoints for communication with MES

+++ src/scheduler/mod.rs
@@ -10,7 +10,7 @@ use crate::{
     scheduler::handlers::{blueprint_handler::ItemBlueprint, order_handler},
 };
 
-pub static CURRENT_DATE: Lazy<RwLock<i32>> = Lazy::new(|| RwLock::new(0));
+pub static CURRENT_DATE: Lazy<RwLock<u32>> = Lazy::new(|| RwLock::new(0));
 pub const TIME_IN_DAY: i64 = 60; // in the simulation, 1 day is 60 seconds
 
 pub struct Scheduler {
@@ -63,7 +63,7 @@ impl Scheduler {
                     }
                 };
 
-                match bp.schedule(order.due_date(), current_date) {
+                match bp.schedule(order.due_date(), current_date as i32) {
                     Ok(_) => Some(bp),
                     Err(e) => {
                         tracing::error!("{:?}", e);
